Added canvas component, moved home and play components to view folder
