+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "classnames": "^2.2.6",
     "mobx": "^3.1.15",
     "mobx-react": "~4.1.0",
     "mobx-state-tree": "^1.3.1",
