+++ src/components/buttons/ImperfectButton.js
@@ -1,20 +1,18 @@
 import React, { Component } from 'react';
 import './ImperfectButton.css';
+import classNames from 'classnames';
 
 import PropTypes from 'prop-types';
 
 class ImperfectButton extends Component {
-	constructor(props) {
-		super(props);
-		this.buttonText = this.props.buttonText;
-	}
 
 	render() {
+
 		return (
-			<div className='button' 
+			<div className='button'
 				onClick={this.props.onClick}
 			>
-				{this.buttonText}
+				{this.props.buttonText}
 			</div>
 		);
 	}
@@ -24,5 +22,6 @@ export default ImperfectButton;
 
 ImperfectButton.propTypes = {
 	buttonText : PropTypes.string,
+	disabled: PropTypes.bool,
 	onClick : PropTypes.func
 }
