+++ src/components/buttons/ImperfectButton.scss
@@ -11,4 +11,5 @@
     text-decoration: none;
     text-align: center;
     cursor: pointer;
+    margin: auto;
 }
