+++ src/components/canvas/Canvas.scss
@@ -1,0 +1,42 @@
+.canvas {
+    background-color: white;
+	height: 90%;
+    width:90%;
+    border-color: black;
+    border-style: solid;
+    border-width: 7px;
+    border-radius: 15px 225px 15px 225px/225px 15px 225px 15px;
+    
+}
+
+.instructions {
+    background-color: rgba(0,0,0,.5);
+    width: 50%;
+    height: auto;
+    cursor: crosshair;
+    border: 2px solid white;
+    padding: 20px;
+    margin: auto;
+    align-self: center;
+    align-content: center;
+    align-items: center;
+    text-align: center; 
+    border-color: black;
+    border-style: solid;
+    border-width: 7px;
+    border-radius: 105px 235px 11px 25px/25px 41px 215px 95px;
+    font-family: 'Amatic SC', cursive;
+    font-size: 65px;
+
+    .button {
+        align-self: center;
+    }
+}
+
+.notActive {
+    cursor: not-allowed;
+}
+
+.disabled {
+    display: none;
+}
