+++ src/components/views/play/Play.js
@@ -1,0 +1,17 @@
+import React, { Component } from 'react';
+import Canvas from '../../canvas/Canvas';
+import './Play.css';
+
+class Play extends Component {
+
+	
+	render() {
+		return (
+			<div className='play'>
+				<Canvas />
+			</div>
+		);
+	}
+}
+
+export default Play;
