+++ src/components/views/play/Play.scss
@@ -13,9 +13,3 @@
     width: 10%;
   }
 }
-
-.canvas {
-	background-color: white;
-	height: 70%;
-	width:50%;
-}
